Add multi-node capability to HotFile
A simulation can now be restarted from a multi-node simulation. This is
possible with the same or any other node configuration. To start from a
multi-node simulation any hotfile needs to be specified and the others
will be loaded automatically. This assumes that the filename is of the
form hot_nX.Y_ZZZZZ.bin, where X is any rank, Y the number of ranks of
the previous simulation and ZZZZZ the index of the restart.