Fix SPH core utils inclusion in post_process_kernel
With the refactoring of the post-process kernels, W and F were made
available by including sph_core_utils.cuh. However, this would have
required setting all the kernel constants and coefficients, which was
not done. Solve by `using namespace cuforces;` instead of including
sph_core_utils, so that we can take advantage of the coefficients being
set in forces.

This fixes the nonsense (typically null) coming out of the postprocess
filters.