Debug flag to inspect the particles before forces computation
When debugging some formulations, we may want to look at the state of
the particles right before the forces get computed, which may be
different from the state of the particles at the end of the integration
step (e.g. because of preprocessing filters).

When this flag is enabled, the particle state will be dumped before each
forces computation (in additional to standard saves).