VTKWriter: append current step when saving preforce state
If the write_flag for the current save is an integrator step, we're
doing an intermediate dump for the inspect_preforce debug option, and
the current time could be the same as the time for the previous save,
which would end up confusing in Paraview due to multiple .vtu files
mapping to the same timestep.

Solve this by 'shifting' the time forward by a very small amount
(related to the integration step) in these cases.