Fix: particle mass is stored in m_center but 1. undocumented 2. overwritten
1. Add minor comment in member doxy doc
2. This happened esp. in setEulerParameters() in Cube and Cylinder

Bug symptons: the body does not feel the effect of the fluid particles at
all if setParticleMass*() is called before any rotation (even a rotation
by null angles, which calls SetEulerParameters() anyway).

This bug could potentially affect cylinder version of shiplock problem,
golours, and any problem with a floating cube or cylinder the particle mass
of which is set before rotating it.