Introduce self-explicative Problem::initializeObjectJoints()
Since problems can access m_bodies_physical_system, the callback is
actually quite powerful, but its aim is to initialize and add joints
(ChLinks in Chrono speak) to the physical world. Since Chrono offers
a number of different joint types, each with specific parameters,
making specific XProblem call to encapsulate them all was unfeasable.

The callback is of course called after fill_parts(), in which Chrono
bodies are initialized.

For a list of the ChLink types see:
http://api.chrono.projectchrono.org/classchrono_1_1_ch_link.html