Fix use of MPI with non-default compiler
In systems where the default CXX compiler is too new for the CUDA
version, we usually install an older, CUDA-supported version of CXX and
use that (e.g. by setting CXX in Makefile.local). However, the
NetworkManager gets compiled with `mpicxx` that will normally still use
the default compiler, whose standard library includes might be
incompatible with the ones of the CXX version we use.

For example, I have g++ 6 as default compiler, and g++ 4.9 is used for
CUDA. Trying to compile GPUSPH gives an error about a missing
include_next for stdlib.h

The solution is to specify for `mpicxx` the same compiler that is being
used for everything else; this can be done by setting an environment
variable when calling `mpicxx`. The variable name depends on the MPI
implementation, so we set both the OpenMPI and MVAPICH one.