Use neighbor search radius for cell size
When the neighbor search radius is (significantly) larger than the
influence radius, it is important that the cell size it at least as
large as the neighbor search, or some potential neighbors might be
missed (particularly for particles near one of the cell sides).

As a side-effect, geometries that would generate too many cells can now
be coped with by setting a larger neighbor search radius.