Allocate pinned memory from GPUWorker
Move the allocation (and deallocation) of pinned memory for CellStarts
and CellEnds from GPUSPH to GPUWorker.

This eliminates the last of the cuda*() calls done from GPUSPH, which
should avoid spurious creations of CUDA contexts that would use device 0
regardless of whether it'll be used by GPUSPH or not. (This is
particularly important becase if devices are in exclusive mode, the
unused context would needlessly block a GPU.)