Update to the new Chrono API
These changes allow us to build against the current (at the time of
commit) HEAD of the develop branch of the Chrono project, with hash
e3be7dee549a6574bc1f3d20823ca8d03850890b.

Note that the changes are incompatible with the old API (around commit
45abc4ff48725ea6baf95d9b8a37075e1c53d440 of the Chrono develop branch),
and we have no sure way of knowing what the user might have, so we can't
support both. Users will have to upgrade to the latest Chrono to use
this version of GPUSPH.