Big simplification of relEulerVel usage
There's no need to make the structure depend on the template
parameters, just the constructor will do. This allows us to make a
simple function (rather than a functor) to handle access to the
relEulerVel field. And since this is always used by adding it to the
standard relVel, we can have the function return the sum directly (if
needed) or just relVel otherwise.