Splitneibs merge: fix duplicate initGamma
The merge resulted in a duplicate command for gamma initialization
(SA_INIT_GAMMA vs INIT_GAMMA), called at different places of the
initialization step.

The distinction was apparently in the dynamic versus quadrature gamma,
but in reality the initialization is the same in both cases (since the
quadrature value is used as initial condition for the dynamic case
anyway). So just call SA_INIT_GAMMA unconditionally, and remove
INIT_GAMMA. Take the opportunity to bring in some minor cleanups that
were in master (such as the use of `fabsf` etc instead of `fabs` etc).