Reorganize overflow checks during neighbor list construction
Refactor the checks for excess neighbors of each type, and use it both
to avoid overflowing the neighborlist and during the final marker
placement to check for overflow and store the id and data of an
overflowing particle. A marker is placed also at the end of the vertex
list.

For the statistics, handle both vertex and non-vertex particles,
separately.