Reintroduce CFL for dynamic gamma
When gamma is being computed dynamically, a time-stepping condition
based on the relative velocity of the fluid the boundary is needed. This
is computed during the FLUID/BOUNDARY interaction, and reduced in the
force finalization step.

Following the master rather than splitneibs branch, this is computed
both when density sum is enabled and when it is not.