Clear up buffer handling for density diffusion
In the density sum case, density diffusion is now applied after
standard integration, but we should still use the standard approach of
putting the “current” state in READ when calling kernels. So split the
computation of the density diffusion term from the application of it,
and wrap them in appropriate buffer swaps to ensure the correct buffers
are being read at every step.