Rewrite fmax reduction for two-step process
Optimal reductions on GPU only take two steps: one to produce a small
number of partial reductions (one per block, with each block processing
as large a number of items as necessary), and one to reduce the partial
reductions.

The two-step approach is not only faster, but also safer, as it only
recycles tempCfl during the second step, which only uses 1 block, and
thus cannot write to it before reading all the values, which was
something at least theoretically possible with the old iterative
approach.

Hence switch to it,with additional constraints, such as the number of
elements to reduce always being a multiple of 4, to maximize throughput
during the initial phase of each step of the reduction. No difference
in results have been observed.