Restore use of fluid sound speed for dt computation
When particle's density grow larger, the sound speed grows larger as
well, and this should be taken into account in the computation of the
adaptive timestep. This was done in master, and removed in
splitneibs, but my experiments show that simulations are more stable if
this is kept, so let's put it back in.