Get rid of dgamdt
Since gamma is now integrated using the integral formulation instead of
the differential continuity equation, there's no need for dgamdt
anymore. This is true even in the open boundary case, since the only
additional contributions pertain the density integration, and in our
formulation they are computed a posteriori after the position and
velocity integration.