Introduce and use enable_if_t
It's a type template equivalent to typename std::enable_if<B, T>::type,
but much shorter to write. It's defined in C++14, but can be defined on
C++11 (which we now support), so make use of it, as it simplifies a lot
of expressions.

For the ones that cannot be simplified that trivially (and particularly
when trying to detect if a struct has a specific member or not), wrap
the detector generation into a macro, and use constexpr boolean
functions instead of requiring ::value.