Restore k-epsilon functionality
The biggest issue was the need to carry over information across split
kernel launches in forces, and the most expensive part of this is the
strain rate tensor information. We recycle the BUFFER_TAU symmetric
tensor which is designed just for this.

(And to be able to do this, we symmetrize the velocity gradient during
computation, to avoid having to store 9 components).