Add missing gamma swap
In the dynamic gamma case, the forces kernel computes the new gamma and
grad gamma, so BUFFER_GRADGAMMA is placed in the WRITE position before
running the kernel, and then swapped back to the READ position on
completion.

However, the swap-back was missing after the kernel invokation of the
second integration step. This was observable even in the simplest of
problems, with very large values of gamma appearing in clusters of
vertex and boundary particles after sorting (a typical phenomenon when
the mismatched, i.e. unsorted buffer is saved).