Improve density diffusion with open boundaries
The reference paper avoids PT_BOUNDARY contributions from solid walls,
but doesn't say anything about their contributions in the open boundary
case. In master they _do_ contribute, and this does improve the
situation, therefore for the time being we enable PT_BOUNDARY
contributions in the open boundary case.

To do this we have to read boundary elements from the correct array, so
make sure we swap and swap back as needed.