Get rid of MultiBufferList iterators
The iterators were based in preparation for the possibility to have
more than two copies per buffer, but that mechanism is going to be
replaced by a better structured approach (ParticleSystem with States).

A ParticleSystem::State will be essentially a single BufferList,
so we just pass BufferLists around now.

The only part that _may_ need checking is the way we use to find the
“other” BufferList index for multi-device transers.