New ID generation algorithm
We count the number of open boundary vertices, and assign each a unique,
sequential new “next ID”, which will be used as the ID of the next
generated particles. When a new particle is generated, the “next ID” of
the generating vertex is incremented by the number of open boundary
vertices.