Clear buffer lists correctly
With the switch to the simpler MultiBufferList without BufferList
vectors, we were not clearing the (two) BufferLists correctly, which
resulted in the corresponding buffer deletion happening in the
destructor which was being invoked at the destruction of the GPUWorker.

However, GPUWorker destruction happens after a call to cuaDeviceReset()
resulting in invalid calls to cudaFree().

Solve by clearing the BufferLists correctly.