Start coalescing VTKWriter header writing
The VTKWriter currently suffers from a split header/data writing, which
requires some care when adding buffers to be dumped. Ideally, the split
writing should be transparent to the developers, which should only
specify, for each array to be written, any optional transformation to
be done on the data.

In preparation for this, we start to unify header writing, starting by
hiding the difference between scalar and vector arrays, and providing a
simple method to set the header for separate .xyz and .w parts of
float4 data types.