VTKWriter: coalesce header and data writing API
A single call now takes care of writing the header and writing the
data. Actual data writing is delayed by putting the relevant
instructions into a lambda and storing a vector of them. A separate
call will them proceed to run all lambdas, in the sequence they were
stored, thereby dumping the arrays in the sequence the metadata was
added.