Only apply rotations to moving bodies
When update_normals() was refactored, it was applied unconditionally for
all particles, to ensure that non-rotating normals were copied over to
the new arrays.

However, this moved update_normals() out of the MOVING(pdata.info)
condition, thus all boundary and vertex particles tried to apply
rotations, which caused fixed boundaries to apply the rotation of the
first moving object.

Solve by checking for moving object before applying the rotation.