GPUSPH_DEVICE environment to override default device(s)
If the GPUSPH_DEVICE environment variable is set, it will be parsed
as a comma-separated list of integers to be used as device indices.
This is used as the default device list if no `--device` command-line
option is provided. If `--device` is specified and GPUSPH_DEVICE is not
set, we still fall back to using device 0.