Enable DTADAPT by default and remove flags<> override
The changed introduced in 420fddf466f671ced1afff3a134891a7a68a6c27 by
Alexis effectively broke all the problems that were designed around
having adaptive time-stepping enabled, forcing them to use flags<>
instead of add_flags<>. The issue with this approach is that it is not
future-proof: if a new flag needs to be introduced that “makes sense
enabled by default”, then all problems need to be adjusted again
(example of this was the discussed ENABLE_DYNAMICS flag to separate
thermal problems from dynamics problem and allow independent activation
of each).

We resolve this by:

* removing the flags<> override; individual (sets of) flags must now be
enabled/disabled explicitly by the problem using add_flags<> and
disable_flags<>;
* ENABLE_DTADAPT by default; to preserve compatibility with the previous
state (introduced by Alexis' change) we disable it in all problems that
didn't enable it explicitly;
* on a subsequent commit, we will actually remove the disabling of the
adaptive dt, since the problems actually expect it to be enabled anyway,
and it not being enabled was more of a temporary workaround from the
splitneibs work.