Switch from unstructured grid to polydata
This allows us to get rid of the cell type support array, since vertices
are set in their own Verts element instead of having to specify for each
cell that it's a vertex.

The data savings are minimal (a single byte per particle, which amounts
to less than 2% even for something as simple as DamBreak3D), but there's
no reason to not take them anyway.