Allow runtime selection for framework options from simple value
This avoids the complexity of the manual switches and repeated
conditionals. The implementation builds on the compile-time knowledge
of the option value range, comparing the runtime value with each of
the possible values until one is found or an error is thrown.

If not all possible values are acceptable, individual conditions still
need to be used.