Preliminary draft of command traits
Similarly to buffer traits, the idea here is to allow a unified
definition of the commands and their usage patterns, in this case the
buffer access patterns.

The current definitions are tentative, obtained from a quick analysis of
the existing acces, and will need to be fine-tuned. In particular, we
need better semantics to differentiate write vs update usage, and the
access to both “old” and “new” copies of the buffers in read mode.