Reset nextIDs for generated particles
With open boundaries, it can happen after a lot of generations and
destructions that a new fluid particle is generated on an index which
at some point in the simulation was occupied by some other particle,
resulting in stale data being used for the new particle.

While we reset most of the arrays, we do not reset next ID (because it
would not be used by the fluid particle anyway), potentially resulting
in generated fluid particles getting a non-default next ID due to stale
vertex data at the given location.

Ensure that the value is properly reset, and leave a TODO for the
remaining un-reset particle property buffers.