Multi-state buffers should not be accessible for writing
Except for some special circumstances at the end of the time-step (e.g.
post processing or particle generation and destruction for open
boundaries) buffers shared between states should be read-only, to
prevent an update to invalidate the previous state too.

An intentionally verbose way to override the check is provided for the
few cases when the assumption is intentionall violated.