Preliminary commands for state manipulation
We introduce three new commands to initialize, release and rename a
particle system state, as well as one to share a buffer between states.
Since we are using strings to represent states, the doCommand() syntax
needs to be (temporarily?) extended.

Only minimal uses are currently introduced, further changes will be
introduced progressively command-by-command.