Fix SA_BOUNDARY preventing compilation of non-WENDLAND kernels
Avoid SA_BOUNDARY-specific kernels from being compiled at all when
SA_BOUNDARY is not in use, otherwise compilation will fail with
non-WENDLAND kernels due to the missing implementations of gradGamma<>
in device code.

To achieve thiss, all SA-specific calls are wrapped in a SFINAE
implementation, with the public method calling it, and the non-SA
variant throwing with a runtime error (it shouldn't be called ever
anyway).