Compute maximum kinematic viscosity during CALC_VISC
In the adaptive time-stepping case, the maximum kinematic viscosity
isn't fixed, but depends on the computed viscosity. Reduce the computed
viscosities in CALC_VISC, and store it for usage by dtreduce().

Take the opportunity to move the Newtonian maximum viscosity and the
sound speed computation to uploadConstants, so that they are only
computed once.