Allow validating BufferList get* methods
During this transition phase, it's not always easy to see when a buffer
was forgotten from the state selection. Introduce a member variable that
can be used to verify if a buffer requested device-side was actually
inserted in the list.

The check is off by default, and once enabled it cannot be disabled
(the buffer lists we apply it to are temporary anyway).

When the migration is complete, and all parameter structures are
redesigned to only fetch the necessary buffers, the check will be
permanently enabled.