ParticleSystem: add state_subset_existing()
An alternative to state_subset that doesn't automatically add new
buffers from the pool if they are not present in the particle system
state. The distinction is important because in general we want to use
the growing version for BufferLists used for writing, and this
non-growing version for BufferLists only used for reading.