Use CommandStruct to handle commands
The global next command is now a CommandStruct, instead of being split
across separate members of GlobalData. Dispatch uses a copy of the
global next command, which should make it easier in the future to extend
this to a full-fledged command queue. All runCommand instances, and the
associated functions, are adapted to pass around a CommandStruct const&.

doCommand variability is simplified to the case no-state, one-state,
two-states. Also take the opportunity to fix a MOVE_STATE_BUFFERS
invokation with a single state parameter, that used to work for some odd
reason due to leftover information from previous commands.