Fix gamma CFL issues with striping
The gamma CFL buffer wasn't working correctly with striping enabled,
due to the intermediate split forces kernel calls mixing up with the
reductions. Solve this by making it large enough to account both uses:
accumulation during the split-kernels, and the reduction at the end.

The subsection used for reductions computes its own offset by taking the
standard cflOffset and adding an additional offset equal to the number
of particles (rounded up to 4 for the requirements we have for the
efficient reduction).