Call HotWriter after the next buildneibs
The HotWriter now uses the same logic as all other writers to claim it
needs to write, without extra logic based on the buildneibs frequency.
Instead, when the HotWriter makes a request for a write, in contrast to
other writers this is set as pending, and automatically satisfied on the
next buildneibs.

The main difference compared to the previous logic is that now we write
on the next buildneibs even if it's not a multiple of the frequency
(e.g. because it was triggered by new particles). Delaying the hotwrite
until the next buildneibs is necessary to ensure that resumes are
always correct even in the multi-device case: without this fix, a
particle changing cell at a device boundary at the iteration of the
hotwrite would be lost on resume (this can be seen e.g. with the Seiche
problem before this fix).