Stabilize gridOffset computation in calcHash
Under extreme circumstances, the gridOffset computation could be
unstable, leading to particles “vibrating” between cells even if their
position would not change.

(The most impressive thing in all this is that the circumstances
can indeed present themselves, as seen e.g. when testing the
resumability of Bubble before this patch.)