Expand CommandStruct to include timestep
At different phases we use different dt values (mostly, either dt itself
or dt/2). To handle this, each CommandStruct now has an associated dt()
method that takes a GlobalData and returns the dt to be used for that
command.

This is an essential step in the direction of not needing to pass the
step flag to the commands themselves.