Fallback to Compute Capability 3.0 (Kepler)
Support for Compute Capability 2.0 (Fermi) was deprecated in version 8
of the CUDA SDK, and has been removed since version 9.

While the minimum requirement for the upcoming version 5 of GPUSPH will
be CUDA 8, let's look forward and set CC 3.0 as the default (i.e. if the
CC could not be detected).