Rework dependency generation
Instead of individual massive files for CPU and GPU sources, use
individual files for each, to improve parallelism.

The idea is inspired by the corresponding changes for the WSL branch,
but with a rather different impementation: instead of transforming the
output to define dependencies for both the object and .d file itself, we
only associate the dependency chain to the .o file, and let the object
file depend on the .d file (if dependencies change, the .d file will
change too, and so will the .o file).

Additionally, the .d file itself is generated during the compilation of
the .o file, since either has to change when the dependencies change.