More target vs executable fixes
Since we produce native Windows executable (that need the .exe
extension), our primary targets should be things such as GPUSPH.exe or
DamBreak3D.exe; however, we also want to be able to run `make` without
specifying the extension (`make DamBreak3D`).

The solution is to provide both full-executable-name targets that do the
main job (including both debug suffix and extension, as needed), and a
separate ‘basename’ target if this is different from the full exec name.