Neighbor iterator fixes and improvements
The multi-type neighbor iterator was using the wrong logic when crossing
from one type to the next: it was still using the old value of
current_type rather than the updated one after reaching the end of the
previous type. When switching from FLUID to BOUNDARY, or from BOUNDARY
to VERTEX, this would lead to the search going in the wrong direction.

The fix is simply to use the current_type data member rather than
passing it as an argument to try_next.

While we're at it, also expose the neighborlist offset index, which
allows iterators to be used to address other data structures which
are addressable like the neighbor list (e.g. sparse matrices with one
element per neighbor).