Try to kill all MPI processes cleanly
When a process aborts, broadcast a specific signal that other processes
can catch to determine if they have to die earlier. Note that this
doesn't guarantee that all processes will catch it in time, so we still
need the MPI_Abort.

(A possible solution would be to have a separate thread in the
multi-node case that simply loops to check if the message has been
sent, but I'm not sure it'd be more clean.)