Rename Problem to ProblemCore
The first step in the cleaning up and future-proofing the Problem
interface consists of the separation of the core (the old Problem class,
now renamed to ProblemCore) from the user-facing interface (e.g.
XProblem). This is now renamed to Problem, will be versioned, and will
allow multiple APIs to coexist without having to invent names such as
XProblem, YProblem, ZProblem etc.