Fix --debug neibs for non-SA and non-fluid
The check for the neighbor search was wrong for boundary particles, but
simply removing it (and trusting the break, which should be possible)
will fail on non-SA boundary, because there's no room for vertex
particles. Then again, we only need to check vertex neighbors if we have
SA, so simply tune the last particle type checked based on the boundary
model.