Problem API 1: fix initialization with default values
When r0 is not set, setParticleMassByDensity() would fail because the
fallback value (= m_deltap) is not used yet. Work around this with a
wrapper function. (The solution might fail if r0 gets set _after_ the
first call to the wrapper function, but we don't have a way to detect
that.)