Initial infrastructure for saving nodes
Defined a way to select nodes to be written and manage writing from problem.
Actual writer is still missing.
Note: We call this a writer but it is a different concept. We don't have
to dump any data from device, nodes are on host, then we don't share
any code from other writers.

TODO: we defined a GT_FEA_WRITE in analogy to GT_FEA_FORCE. In the
future we could make a unique type (i.e. GT_FEA_SELECT) to select nodes,
and then specify what to do for these nodes.