Improve parsing of options with commands and conjunctions
If we want to e.g. set the periodicity to PERIODIC_XY from the command
line, the standard parsing method isn't sufficient, because the string
representation for the option is "X and Y".

To solve this, we remove conjunctions and commas (optionally replacing
them with _) when checking for string validity.