DEM load/store improvements
There is a discrepancy between the DEM formats we support: some of them
provide vertex-based information, while others provide cell-based
information.

To clarify the difference between the two, consider the case of
a rectangle that is 2m wide (east-west) and 1m tall (south-north),
with a corner at 0 (DEM extents: [0, 2] by [0, 1]), sampled at
a resolution of 0.1m, resuling in the rectangle being
divided into 20x10 squares.

For a cell-based DEM storage, we have the information about the height
of each of these squares (cells), so we have 20 columns and 10 rows of
data. Point-wise, the height information associated with a cell would be
centered in the cell itself: e.g., for the first cell it would be at
coordinates (0.05, 0.05).

For a vertex-based DEM, we instead store information about the height at
the vertices of each cell: we would therefore have information about the
points at (0, 0), (0, 0.1), … (0.1, 0), (0.1, 0.1), …
From edge to edge, this requires us to have 21 columns and 11 rows of
data.

Internally, GPUSPH handles DEM interpolation assuming vertex-based
information (edge-to-edge). This means that XYZ and VTK structured grids
format (which are vertex-based) can be used “natively”,
but it requires special handling for the ASCII grid formats, which are
cell-based.

Historically, GPUSPH simply “reinterpreted” the ASCII grid formats as
being vertex-based as well, stretching the data points so that they
mapped starting from the edges, and thus effectively reducing the
resolution by N/(N-1) (with N being the number of rows/columns) in each
direction.

While this isn't a big issue in most cases (with hundreds of
rows/columns, the discrepancy is of less than 1%), a user may prefer to
have a stricter interpretation of the data.

To this end, this patch introduces a DEM_FMT_ASCII_STRICT, that
maps the data from cell-based ASCII formats to their proper coordinates
(cell centers). Since the internal representation is still vertex-based,
however, in this case we reduce the domain by half a cell in each
direction (so that the edges of the area covered by the DEM correspond
to the location of the center of the edge cells).

While we're at it, we also fix the VTK writer avoiding a spurious offset
in the computation of the DEM point coordinates.