Take into account DEM offset and world origin
The DEM offset was used inconsistently on the host side (sometimes
assumed to be 0, some other times used as is (see issue #61 on the
public repository), and completely ignored on the device side,

Host-side, the issue is resolved systematically by separating the
computation of the DEM interpolation and particle/DEM distance into an
“internal” version (that takes as input xy coordinates _relative to the
DEM origin_) and an “external” version (that takes as input global
coordinates).

Device-side, the issue is resolved by introducing a “fixup” for the
DemPos conversion (from grid + lpos to DEM coordinates) that takes into
consideration the discrepancy between the DEM and world origin (this is
already corrected for the DEM resolution scale in the two coordinates,
to avoid useless repeated computations).