Rewrite reorderDataAndFindCellStartDevice to not use textures
The changes go much deeper, by implementing a long-pending change that
refactors the data reorder into independent chunks, one for each buffer,
which are all managed by a conditional structure.

The restructuring allows us to validate buffer access to the buffer
lists, meaning that trying to get into the list a buffer that doesn't
exist will lead to an error. This requires special handling of the
buffers that are not initialized (and thus not sorted) during the first
reorder (e.g. dummy velocity and gamma gradient).