Associate cudaTextureObjects to each array in a CUDABuffer
This is not done for all CUDABuffers because some of them have types
(typically, float3 as in BUFFER_VORTICITY and BUFFER_DTKTE) that cannot
be mapped to textures directly.

The solution is to split the texture object handling to a separate
class, and introduce a mock-up of that same class that doesn't actually
do anything. The CUDABufferImplementaiton then subclasses one or the
other depending on the (size of) the type of the elements in the buffer.