Introduce neibs_interaction_params
This is a structure template based on the boundary type assembled from
the neibs_list_params (to iterate over neighbors), the vel_wrapper (to
get the density/volume) and, in the SA_BOUNDARY case, also the
sa_boundary_params (gamma gradient, boundary elements and vertex
positions).

This new structure will be used by all (or at least most) kernels with a
particle/particle interaction.