Remove static demTex in favor of bindless texture object
The approach used is not the cleanest (it would be nice to handle this
through the buffer system too), but it'll do for the time being.

Since demTex is not global anymore, we need to pass it around as kernel
argument, which requires a refactoring to the DEM contribution into a
SFINAEd specialization. While we're at it, we SFINAE the plane
contribution too.