Move phys wrappers to phys core
forces_kernel.def defines a plethora of wrappers to access / manipulate
physical information of particles from the particle data structure
template. They are defined here because this is where they have been
used so far. As we extend support for ghost particles to other kernels,
these same wrappers and auxiliary functions need to be made accessible,
so we need to move them to more general namespaces.