fixup: nodes handling for cylinders
The generic nodes handling (Hexagons, shells and cables) needs to be
restructured. In fact, now that the nodes can be reused among geometries
the approach of passing onlu few nodes and obtaining the others by
adding 1 is not valid anymore.
This is a temporary fixup not contemplating Hexagons. Meditate if we
should deprecate hexagons.