Problem API 1: distinction between inner and outer borders
In multi-layer problems, the distinction is important as it determines
the FillIn direction. This allow placing internal obstacles without
complex user-side computations about the size and offset of the
inner obstacles.

The (now legacy) FT_BORDER value is kept as an alias for FT_OUTER_BORDER
for backwards compatibility reason.