XYZReader: more robust reading
XYZ files are read in two phases: during the first one we determine the
number of particles by checking how many lines we can read. During the
second phase, we read the actual particle data (and ignore anything else
on the line).

This step used to be done by reading triplets and setting the
corresponding array element until EOF was reached. The issue is that EOF
detection in stream input triggers on the first read that gets to the
end of file, which means that if the XYZ ended with a newline we would
attempt to read one triplet more than necessary, and assign to an
out-of-bound ReadParticle element, before finishing the loop.

However, during the second phase we don't actually need to check for EOF
at all, since we know how many particles we expect. So fix the
off-by-one error by simply looping npart times.

We additionally mark the input stream to throw an exception of failed
conversion, to ensure we're reading the correct data, and print a
warning if we're reading the file a second time (detected by the buffer
being already allocated).

This fixes GitHub issue #71.