Remove currently unused dt variable in MOVE_BODIES
As explained in the following TODO, we use the dt from gdata instead of
the one from the step. We might want to consider using the step
information instead, but this needs to be done in concert with the
test cases to ensure that the correct time step is being used at all
times.