Help clang loop unroller
The gamma quadrature constants are “true” constants. They are in device
__constant__ memory purely as a matter of practicality. However, this
prevents clang from unrolling the quadrature loops, as the bounds are
not known exactly.

The solution in clang is to mark the constant arrays constexpr, but nvcc
does not support this (they could be turned into consexpr __device__
functions, but that's rather cumbersome without at least C++14 support),
so we define a macro that uses __constant__ or constexpr depending on
compiler.