RTInstability: simplify initialization logic
Most of the logic is shared between fluid and boundary particles. The
only exception is the light fluid particles. So reduce the conditionals
and unify most of the code, specializing only the light fluid treatment.

Also use set_fluid_num() since that's the only field we're changing in
the particle info.