DEM: automatic handling of half-pixel correction
CUDA textures are pixel-centered, so that the value for the pixel at
row r column c of the original array is obtained fetching from (c+0.5f,
r+0.5f) rather than (c, r). We were adding this offset manually at
DemPos calculation time, but this requires subtracting the offset again
when interpolaing manually.

Since this offset is an impementation detail, manage the shift
internally in fetchDem, adding it when using textures and avoiding doing
a +0.5f (in DemPos) - 0.5f (in fetchDem) when textures are disabled.