Improve hotstart with moving bodies
The idea is to store (and restore) the linear and angular acceleration
of the body. The original patch was provided by @JoJo on the Discourse
forum (aka ZhenxiZhao on GitHub).

However, since the new storge requirements overflow the original
served sppace for objects in VERSION_1 files, I changed it to do a
version bump on the HotFile.