Avoid conflict between type as typedef and function
An interesting side-effect of having the kernel functor derive from the
struct params is that some internal struct members get exposed to the
kernel body (operator()). This is usually not an issue, but a couple of
structures have a `using type = sometype` directive that overrides
the type() function that extracts the particle type.

It could be argued that the type() function is not named in the best
possible way, and it should be renamed to something longer and less
prone to conflict, but in this case it's simpler (and possibly more
resembling of the std typedefs) to rename the type member to
element_type.