Makefile: fix #include in shell
Behind commit 3e704aceb7c4c28d4b94a879a07b2a4b73f2267c there was a
strange incosistency in behavior between machines, possibly related to
the GNU make version.

A solution that seems to work correctly is the use of \043 to get a # in
printf without interfering with the shell or Makefile parsing (whoever
is responsible for the difference in behavior of # inside a quoted
string inside a shell call).