Make debug flags a global object
Instead of passing it through GlobalData, use a global g_debug instance.
We were already doing some selective passthrough for the BUFFER_ACCESS
debugging, but the flags are being used all through the code,
and having to pass them around individually is a nuisance (especially
when we need to change several function signatures to achieve that,
like we do with the one that get to the kernels), so let's just go
global.