Expunge leftover as_<type> use and definitions
Some compilers (hello, nvcc) don't handle them as one would expect,
leading to inefficient code, so let's just remove them altogether
to spare us the temptation.

On host we can just use make_* instead, both because host compilers are
generally smarter, and because it's not the hot path anyway (so even if
the make_* is slower, who cares).

On the device, its usage has been replaced over the course of the past
commits with the use of specific structures isolating the space and 4th
component of each type. This way we don't have to worry about the
possible inefficiency of make_* and the known inefficiency of as_*

(On my machine, DamBreak3D went from 11 to 13 MIPPS.)